Moved config files. Changed ManipulationController implementation
