--- src/main/java/com/example/hello_world/web/service/ProfileService.java
@@ -1,16 +1,0 @@
-package com.example.hello_world.web.service;
-
-public class ProfileService {
-
-
-    private String profile;
-
-    public ProfileService(String profile) {
-        this.profile = profile;
-    }
-
-    @Override
-    public String toString() {
-        return "ProfileService{" + "profile='" + profile + '\'' + '}';
-    }
-}
